requirements.txt
pdf2image
opencv-python
pillow
numpy
tqdm